refactor: mudança de tipo do ID em Store e StoreRepository; adição de construtor com argumentos

+++ src/main/java/com/business_management/repository/StoreRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.business_management.domain.Store;
 
 @Repository
-public interface StoreRepository extends JpaRepository<Store, Integer> {
+public interface StoreRepository extends JpaRepository<Store, Long> {
 
 }
